✨ feat(mobile/component): add Content component for displaying various content types
- Introduced a new Content component that encapsulates the display logic for articles, videos, papers, and blogs.
- Updated the Home component to utilize the new Content component, improving code organization and readability.
- Removed redundant type icon and label functions from Home, centralizing this logic within the Content component.

This commit enhances the modularity of the codebase and improves the user interface for content presentation.
